requirements.txt
Django
djangorestframework
Pillow
flake8
mysqlclient